Moving sound-specific files to src/lib
