+++ src/lib/react-sound-plus.js
