+++ src/lib/sound-sprite.js
